+++ regression/reference/struct-cxx/wrapfstruct.f
@@ -919,7 +919,7 @@ contains
     function cstruct_as_class_get_x1(obj) &
             result(SHT_rv)
         use iso_c_binding, only : C_INT
-        class(cstruct_as_class) :: obj
+        class(cstruct_as_class), intent(INOUT) :: obj
         integer(C_INT) :: SHT_rv
         ! splicer begin class.Cstruct_as_class.method.get_x1
         SHT_rv = c_cstruct_as_class_get_x1(obj%cxxmem)
@@ -937,7 +937,7 @@ contains
     ! start cstruct_as_class_set_x1
     subroutine cstruct_as_class_set_x1(obj, val)
         use iso_c_binding, only : C_INT
-        class(cstruct_as_class) :: obj
+        class(cstruct_as_class), intent(INOUT) :: obj
         integer(C_INT), value, intent(IN) :: val
         ! splicer begin class.Cstruct_as_class.method.set_x1
         call c_cstruct_as_class_set_x1(obj%cxxmem, val)
@@ -953,7 +953,7 @@ contains
     function cstruct_as_class_get_y1(obj) &
             result(SHT_rv)
         use iso_c_binding, only : C_INT
-        class(cstruct_as_class) :: obj
+        class(cstruct_as_class), intent(INOUT) :: obj
         integer(C_INT) :: SHT_rv
         ! splicer begin class.Cstruct_as_class.method.get_y1
         SHT_rv = c_cstruct_as_class_get_y1(obj%cxxmem)
@@ -971,7 +971,7 @@ contains
     ! start cstruct_as_class_set_y1
     subroutine cstruct_as_class_set_y1(obj, val)
         use iso_c_binding, only : C_INT
-        class(cstruct_as_class) :: obj
+        class(cstruct_as_class), intent(INOUT) :: obj
         integer(C_INT), value, intent(IN) :: val
         ! splicer begin class.Cstruct_as_class.method.set_y1
         call c_cstruct_as_class_set_y1(obj%cxxmem, val)
@@ -990,7 +990,7 @@ contains
     function cstruct_as_subclass_get_x1(obj) &
             result(SHT_rv)
         use iso_c_binding, only : C_INT
-        class(cstruct_as_subclass) :: obj
+        class(cstruct_as_subclass), intent(INOUT) :: obj
         integer(C_INT) :: SHT_rv
         ! splicer begin class.Cstruct_as_subclass.method.get_x1
         SHT_rv = c_cstruct_as_subclass_get_x1(obj%cxxmem)
@@ -1008,7 +1008,7 @@ contains
     ! start cstruct_as_subclass_set_x1
     subroutine cstruct_as_subclass_set_x1(obj, val)
         use iso_c_binding, only : C_INT
-        class(cstruct_as_subclass) :: obj
+        class(cstruct_as_subclass), intent(INOUT) :: obj
         integer(C_INT), value, intent(IN) :: val
         ! splicer begin class.Cstruct_as_subclass.method.set_x1
         call c_cstruct_as_subclass_set_x1(obj%cxxmem, val)
@@ -1024,7 +1024,7 @@ contains
     function cstruct_as_subclass_get_y1(obj) &
             result(SHT_rv)
         use iso_c_binding, only : C_INT
-        class(cstruct_as_subclass) :: obj
+        class(cstruct_as_subclass), intent(INOUT) :: obj
         integer(C_INT) :: SHT_rv
         ! splicer begin class.Cstruct_as_subclass.method.get_y1
         SHT_rv = c_cstruct_as_subclass_get_y1(obj%cxxmem)
@@ -1042,7 +1042,7 @@ contains
     ! start cstruct_as_subclass_set_y1
     subroutine cstruct_as_subclass_set_y1(obj, val)
         use iso_c_binding, only : C_INT
-        class(cstruct_as_subclass) :: obj
+        class(cstruct_as_subclass), intent(INOUT) :: obj
         integer(C_INT), value, intent(IN) :: val
         ! splicer begin class.Cstruct_as_subclass.method.set_y1
         call c_cstruct_as_subclass_set_y1(obj%cxxmem, val)
@@ -1058,7 +1058,7 @@ contains
     function cstruct_as_subclass_get_z1(obj) &
             result(SHT_rv)
         use iso_c_binding, only : C_INT
-        class(cstruct_as_subclass) :: obj
+        class(cstruct_as_subclass), intent(INOUT) :: obj
         integer(C_INT) :: SHT_rv
         ! splicer begin class.Cstruct_as_subclass.method.get_z1
         SHT_rv = c_cstruct_as_subclass_get_z1(obj%cxxmem)
@@ -1076,7 +1076,7 @@ contains
     ! start cstruct_as_subclass_set_z1
     subroutine cstruct_as_subclass_set_z1(obj, val)
         use iso_c_binding, only : C_INT
-        class(cstruct_as_subclass) :: obj
+        class(cstruct_as_subclass), intent(INOUT) :: obj
         integer(C_INT), value, intent(IN) :: val
         ! splicer begin class.Cstruct_as_subclass.method.set_z1
         call c_cstruct_as_subclass_set_z1(obj%cxxmem, val)