+++ shroud/typemap.py
@@ -153,6 +153,7 @@ class Typemap(object):
         ("cfi_type", "CFI_type_other"),
         ("is_enum", False),
         ("export", False),      # If True, export to YAML file.
+        ("smart_pointer", None), # Used to generated names
         ("__line__", None),
     )
 
@@ -297,7 +298,7 @@ class Typemap(object):
             order = self._keyorder
         else: # class
             # To be used by other libraries which import shadow types. 
-            if self.base in ["shadow", "shared"]:
+            if self.base in ["shadow", "smartptr"]:
                 order = [
                     "base",
                     "wrap_header",
@@ -927,8 +928,19 @@ def default_typemap():
             cxx_type="std::shared_ptr<{cxx_T}>",
             cxx_header="<memory>",
             impl_header=["<memory>"],
-            base="shared_ptr",
-            sgroup="shared_ptr",
+            base="smart_ptr",
+            sgroup="smart_ptr",
+            smart_pointer="shared",
+        ),
+        weak_ptr=Typemap(
+            "std::weak_ptr",
+            ntemplate_args=1,
+            cxx_type="std::weak_ptr<{cxx_T}>",
+            cxx_header="<memory>",
+            impl_header=["<memory>"],
+            base="smart_ptr",
+            sgroup="smart_ptr",
+            smart_pointer="weak",
         ),
         MPI_Comm=Typemap(
             "MPI_Comm",
@@ -955,6 +967,8 @@ def default_typemap():
     del def_types["vector"]
     def_types["std::shared_ptr"] = def_types["shared_ptr"]
     del def_types["shared_ptr"]
+    def_types["std::weak_ptr"] = def_types["weak_ptr"]
+    del def_types["weak_ptr"]
 
     # One typemap for all template parameters.
     type_name = "--template-parameter--"
@@ -1507,7 +1521,7 @@ def return_user_types(typemaps):  # typemaps -> dict
         if ntypemap.name == "--template-parameter--":
             continue
         elif ntypemap.sgroup in [
-                "shadow", "shared", "struct", "template", "enum", "procedure"]:
+                "shadow", "smartptr", "struct", "template", "enum", "procedure"]:
             dct[key] = ntypemap
         elif ntypemap.is_enum:
             dct[key] = ntypemap