+++ docs/releases.rst
@@ -342,6 +342,11 @@ New Features
 * Add format fields *F_name_typedef* and *C_name_typedef* to name a typedef
   for Fortran or C.
 
+* Add intent attribute to a function's ``PASS`` argument. If the C++ function
+  is ``const`` set to ``intent(IN). Otherwise, set to ``intent(INOUT)``.
+
+* Added attribute *+operator(assignment)* to add a Fortran assignment overload.
+
 Fixed
 ^^^^^